requirements.txt
APScheduler==3.6.3
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
idna==2.10
iniconfig==1.0.1
more-itertools==8.5.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.2
python-telegram-bot==13.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
tornado==6.1
tweepy==3.9.0
tzlocal==2.1
urllib3==1.25.10
